@@ -1,12 +1,11 @@
 name: ios_specific_csx
 description: A new flutter plugin project.
-version: 0.0.1+3
+version: 0.1.0
 homepage: https://github.com/KirstenDunst/FlutterPlugins/tree/main/ios_specific_csx
-publish_to: http://dart-pub.brainco.cn:8081
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
