@@ -1,11 +1,17 @@
+//
+//  IosSpecificCsxHealthPlugin.swift
+//  ios_specific_csx
+//
+//  Created by 曹世鑫 on 2021/5/8.
+//
 import Flutter
 import UIKit
 
-public class SwiftIosSpecificCsxPlugin: NSObject, FlutterPlugin {
+public class IosSpecificCsxHealthPlugin: NSObject, FlutterPlugin {
     
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "ios_specific_csx", binaryMessenger: registrar.messenger())
-        let instance = SwiftIosSpecificCsxPlugin()
+        let channel = FlutterMethodChannel(name: "ios_specific_csx_health", binaryMessenger: registrar.messenger())
+        let instance = IosSpecificCsxHealthPlugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
     }
 
@@ -16,15 +22,6 @@ public class SwiftIosSpecificCsxPlugin: NSObject, FlutterPlugin {
             HealthTool.sharedInstance.addHealthMindfulness(startDateInMill: argument["startDate"] as! Int, endDateInMill: argument["endDate"] as! Int) { (success, errorStr) in
                 result(["success":success,"errorDescri":errorStr as Any])
             }
-            break;
-        case "requestHealthAuthority":
-            HealthTool.sharedInstance.requestHealthAuthority { (success, errorStr) in
-                result(["success":success,"errorDescri":errorStr as Any])
-            }
-            break;
-        case "getHealthAuthorityStatus":
-            let answer = HealthTool.sharedInstance.getHealthAuthorityStatus(subclassificationIndex: call.arguments as! Int).rawValue
-            result(answer)
             break;
         case "addHealthStature":
             let argument = call.arguments as! NSDictionary;
@@ -78,18 +75,23 @@ public class SwiftIosSpecificCsxPlugin: NSObject, FlutterPlugin {
             let argument = call.arguments as! NSDictionary;
             HealthTool.sharedInstance.addHealthHeartRate(heartRate: argument["heartRate"] as! Double,startDateInMill:argument["startDate"] as? Int, endDateInMill:argument["endDate"] as? Int) { (success, errorStr) in
                 result(["success":success,"errorDescri":errorStr as Any])
+            }
             break;
-        case "isHealthDataAvailable":
-            let answer = HealthTool.sharedInstance.isHealthDataAvailable()
+        case "requestHealthAuthority":
+            HealthTool.sharedInstance.requestHealthAuthority(subclassificationIndex: call.arguments as! Int) { (success, errorStr) in
+                result(["success":success,"errorDescri":errorStr as Any])
+            }
+            break;
+        case "getHealthAuthorityStatus":
+            let answer = HealthTool.sharedInstance.getHealthAuthorityStatus(subclassificationIndex: call.arguments as! Int).rawValue
             result(answer)
             break;
+        case "isHealthDataAvailable":
+            result(HealthTool.sharedInstance.isHealthDataAvailable())
+            break;
         case "gotoHealthApp":
-            guard let url = URL.init(string: "x-apple-health://app/") else {
-                result(false)
-                fatalError("生成跳转链接失败")
-            }
-            UIApplication.shared.open(url){(success) in
-                result(success)
+            HealthTool.sharedInstance.gotoHealthApp { (success, errorStr) in
+                result(["success":success,"errorDescri":errorStr as Any])
             }
             break;
         default:
@@ -98,3 +100,4 @@ public class SwiftIosSpecificCsxPlugin: NSObject, FlutterPlugin {
        
     }
 }
+
